Split event service unit tests into separate files
